fix(warnings): removed a bunch of warnings from conversion of std::size_t to uint_16_t

+++ Client/Source/System/ControlSystem/ControlSystem.hpp
@@ -43,7 +43,7 @@ class ControlSystem {
 		 * @param velocities Sparse array of all velocity component of all entities
 		 */
 		void operator()(Registry &registry, Sparse_array<component::cposition_t> &positions, Sparse_array<component::cvelocity_t> &velocities, Sparse_array<component::ckeyboard_t> &keyboards, Sparse_array<component::cnetwork_queue_t> &network, Sparse_array<component::cid_of_ship_t> &idOfShip);
-		
+
 		/**
 		 * @brief A method to add a task for the server through the network queue
 		 * 
@@ -52,7 +52,7 @@ class ControlSystem {
 		 * @param network 
 		 * @param idOfShip 
 		 */
-		void addToNetworkQueue(uint16_t x, uint16_t y, Sparse_array<component::cnetwork_queue_t> &network, uint16_t idOfShip);
+		void addToNetworkQueue(float x, float y, Sparse_array<component::cnetwork_queue_t> &network, uint16_t idOfShip);
 
 
 	protected:
